Fix for broken div id change
Yahoo changed one of the divs during the week of 3/3.  This one line update brings it back into conformity. SigFig is an investment advisor of sorts and the term 'promo' leads one to believe that this odd change to the div may be temporary - if so a reversion to the prior may then be possible.