Update yahoo.finance.oquote.xml
Recent changes to the Yahoo finance quote pages have severely broken a number of datatables, including oquote.   The javascript in the above file has been change substantially to again return data.  Minor documentation of the changes is included and contextual examples of the prior oquote javascript are kept to provide background for any future modifications.  The default ticker is set for an option expiring two years in the future (2017).